allow nft upload mp4 in create flow
