+++ src/components/UploadNoStyle/UploadNoStyle.tsx
@@ -1,10 +1,17 @@
 import { t } from '@lingui/macro'
-import { Upload, UploadProps } from 'antd'
+import { Progress, Upload, UploadProps } from 'antd'
 import { RcFile } from 'antd/lib/upload'
 import { FormItemInput } from 'models/formItemInput'
-import { ReactNode, useCallback, useState } from 'react'
+import { ReactNode, useCallback, useContext, useState } from 'react'
 import { emitErrorNotification } from 'utils/notifications'
 import { UploadRequestOption } from 'rc-upload/lib/interface'
+import { CreateButton } from 'components/CreateButton'
+import { UploadOutlined, CloseCircleFilled } from '@ant-design/icons'
+import { RewardImage } from 'components/Create/components/RewardImage'
+import { stopPropagation } from 'react-stop-propagation'
+import { ThemeContext } from 'contexts/themeContext'
+import { JuiceVideoPreview } from 'components/v2v3/JuiceVideoPreview'
+import { useContentType } from 'hooks/ContentType'
 
 interface UploadNoStyleProps
   extends FormItemInput<string | undefined>,
@@ -13,7 +20,9 @@ interface UploadNoStyleProps
       'onChange' | 'showUploadList' | 'children' | 'customRequest'
     > {
   sizeLimit?: number
-  supportedFileTypes?: Set<'image/jpeg' | 'image/png' | 'image/gif'>
+  supportedFileTypes?: Set<
+    'image/jpeg' | 'image/png' | 'image/gif' | 'video/mp4'
+  >
   customRequest?: (options: UploadRequestOption) => Promise<string> | string
   children?: (props: {
     percent: number | undefined
@@ -24,6 +33,10 @@ interface UploadNoStyleProps
 }
 
 export const UploadNoStyle = (props: UploadNoStyleProps) => {
+  const {
+    theme: { colors },
+  } = useContext(ThemeContext)
+
   const [_uploadUrl, _setUploadUrl] = useState<string>()
   const [isUploading, setIsUploading] = useState<boolean>(false)
   const [percent, setPercent] = useState<number | undefined>(undefined)
@@ -32,6 +45,9 @@ export const UploadNoStyle = (props: UploadNoStyleProps) => {
   const setUploadUrl = props.onChange ?? _setUploadUrl
 
   const undo = useCallback(() => setUploadUrl(undefined), [setUploadUrl])
+
+  const contentType = useContentType(uploadUrl)
+  const isVideo = contentType === 'video/mp4'
 
   const handleBeforeUpload = useCallback(
     async (file: RcFile) => {
@@ -88,6 +104,34 @@ export const UploadNoStyle = (props: UploadNoStyleProps) => {
     [props, setUploadUrl],
   )
 
+  const _body = () => {
+    if (isUploading) {
+      return (
+        <div>
+          <Progress
+            width={48}
+            className="h-8 w-8"
+            strokeColor={colors.background.action.primary}
+            type="circle"
+            percent={percent}
+            format={percent => (
+              <div className="text-black dark:text-grey-200">
+                {percent ?? 0}%
+              </div>
+            )}
+          />
+        </div>
+      )
+    }
+    if (uploadUrl === undefined) {
+      return <UploadButton />
+    }
+    if (isVideo) {
+      return <JuiceVideoPreview src={uploadUrl} />
+    }
+    return <UploadedImage imageUrl={uploadUrl} onRemoveImageClicked={undo} />
+  }
+
   return (
     <Upload
       {...props}
@@ -97,9 +141,37 @@ export const UploadNoStyle = (props: UploadNoStyleProps) => {
       customRequest={onCustomRequest}
       showUploadList={false}
     >
-      {props.children
-        ? props.children({ uploadUrl, isUploading, undo, percent })
-        : null}
+      {_body()}
     </Upload>
+  )
+}
+
+const UploadButton = () => {
+  return (
+    <CreateButton icon={<UploadOutlined />} className="h-24 w-full">
+      Upload
+    </CreateButton>
+  )
+}
+
+const UploadedImage = ({
+  imageUrl,
+  onRemoveImageClicked,
+}: {
+  imageUrl: string
+  onRemoveImageClicked?: VoidFunction
+}) => {
+  return (
+    <div className="flex justify-center bg-smoke-200 py-2 dark:bg-slate-600">
+      <div className="relative">
+        <RewardImage className="h-[11.5rem] w-[11.5rem]" src={imageUrl} />
+        <CloseCircleFilled
+          className="absolute top-0 right-0 cursor-pointer text-2xl text-haze-400"
+          // TODO: We require @tailwind base to do this in className, so use style for now
+          style={{ transform: 'translate(50%, -50%)' }}
+          onClick={stopPropagation(onRemoveImageClicked)}
+        />
+      </div>
+    </div>
   )
 }
