+++ src/locales/messages.pot
@@ -1307,6 +1307,9 @@ msgstr ""
 msgid "Fees held"
 msgstr ""
 
+msgid "File"
+msgstr ""
+
 msgid "File empty or corrupt. Try again."
 msgstr ""
 
@@ -1506,9 +1509,6 @@ msgid "If you're unsure if you need to claim, you probably don't."
 msgstr ""
 
 msgid "Image file"
-msgstr ""
-
-msgid "Image will be cropped to a 1:1 square in thumbnail previews on the Juicebox app."
 msgstr ""
 
 msgid "Image will be cropped to a square in thumbnail previews on the Juicebox app."
@@ -1836,6 +1836,9 @@ msgid "NFT mint confirmed"
 msgstr ""
 
 msgid "NFT projects"
+msgstr ""
+
+msgid "NFT will be cropped to a 1:1 square in thumbnail previews on the Juicebox app."
 msgstr ""
 
 msgid "NFTs"
